requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
Flask-Caching==2.1.0
Flask-Login==0.6.3
Pillow==10.3.0
pytest==8.1.1
psycopg2-binary==2.9.11
gunicorn==21.2.0
cloudinary==1.36.0
